refactor: перенес интерфейс TrackPlayer  domain
